some gcc4 warning fixes
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@28207 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/pulse/NormalPulseView.cpp
@@ -238,7 +238,7 @@ NormalPulseView::Pulse()
 		if (Window()->Lock()) {
 			// Set the value of each CPU bar
 			for (int x = 0; x < fCpuCount; x++) {
-				fProgressBars[x]->Set(max_c(0, cpu_times[x] * 100));
+				fProgressBars[x]->Set((int32)max_c(0, cpu_times[x] * 100));
 			}
 
 			Sync();
